API change: move lifecycle related classes to secretsauce.lifecycle
* AttachableViewModel now does not accept onDetachAction in `onAttach` method
It accept any number of actions via `registerDetachAction`(s) methods
