--- SecretSauce/src/main/java/com/byoutline/secretsauce/di/AttachableViewModel.kt
@@ -1,27 +1,0 @@
-package com.byoutline.secretsauce.di
-
-import android.arch.lifecycle.ViewModel
-
-/**
- * [ViewModel] that should be attached and detached to the view.
- */
-abstract class AttachableViewModel<T> : ViewModel() {
-
-    var view: T? = null
-    private var onDetachAction: (() -> Any)? = null
-
-    open fun onAttach(view: T) {
-        onAttach(view, {})
-    }
-
-    protected fun onAttach(view: T, onDetachAction: () -> Any = {}) {
-        this.view = view
-        this.onDetachAction = onDetachAction
-    }
-
-    open fun onDetach() {
-        onDetachAction?.invoke()
-        onDetachAction = null
-        view = null
-    }
-}
