+++ SecretSauce/src/main/java/com/byoutline/secretsauce/lifecycle/AttachableViewModelDetachAction.java
@@ -1,0 +1,7 @@
+package com.byoutline.secretsauce.lifecycle;
+
+
+@FunctionalInterface
+public interface AttachableViewModelDetachAction {
+    void onDetach();
+}
