+++ sampleapplication/src/main/java/com/byoutline/sampleapplication/draweractivity/DrawerActivity.kt
@@ -12,8 +12,8 @@ import com.byoutline.sampleapplication.R
 import com.byoutline.sampleapplication.databinding.ActivityExampleBinding
 import com.byoutline.secretsauce.activities.clearBackStack
 import com.byoutline.secretsauce.activities.showFragment
-import com.byoutline.secretsauce.di.bindContentView
-import com.byoutline.secretsauce.di.lazyViewModel
+import com.byoutline.secretsauce.lifecycle.bindContentView
+import com.byoutline.secretsauce.lifecycle.lazyViewModel
 import dagger.android.AndroidInjector
 import dagger.android.DispatchingAndroidInjector
 import dagger.android.support.HasSupportFragmentInjector
