+++ sampleapplication/src/main/java/com/byoutline/sampleapplication/rx/RxLifecycleActivity.kt
@@ -5,9 +5,9 @@ import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import com.byoutline.sampleapplication.R
 import com.byoutline.sampleapplication.databinding.RxlifecycleActivityBinding
-import com.byoutline.secretsauce.di.AttachableViewModel
-import com.byoutline.secretsauce.di.bindContentView
-import com.byoutline.secretsauce.di.lazyViewModelWithAutoLifecycle
+import com.byoutline.secretsauce.lifecycle.AttachableViewModel
+import com.byoutline.secretsauce.lifecycle.bindContentView
+import com.byoutline.secretsauce.lifecycle.lazyViewModelWithAutoLifecycle
 import java.util.concurrent.TimeUnit
 import io.reactivex.Observable as RxObservable
 
@@ -32,8 +32,7 @@ class RxLifecycleViewModel : AttachableViewModel<Any>() {
     override fun onAttach(view: Any) {
         val disposable = RxObservable.interval(1, TimeUnit.SECONDS)
                 .subscribe { value.set(value.get() + 1) }
-        super.onAttach(view, onDetachAction = {
-            disposable.dispose()
-        })
+        super.onAttach(view)
+        registerDetachAction { disposable.dispose() }
     }
 }
